Fixed major Error caused by hooks usage
